backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-multipart
python-dotenv
psycopg2-binary
google-cloud-storage